TenantsArtisanCommand was removed instead RegisterTenantScheduleTasks was added
